Fjerner opprettelse av bekreftet periode endring og oppretter uttalelse i steden, splitter ut opprettelse av etterlysning i egen tjeneste
