+++ domenetjenester/mottak/src/main/java/no/nav/ung/sak/mottak/dokumentmottak/oppgavebekreftelse/DatoEndringBekreftelseHåndterer.java
@@ -4,16 +4,16 @@ import jakarta.enterprise.context.Dependent;
 import jakarta.inject.Inject;
 import no.nav.k9.oppgave.bekreftelse.Bekreftelse;
 import no.nav.k9.oppgave.bekreftelse.ung.periodeendring.DatoEndring;
-import no.nav.k9.oppgave.bekreftelse.ung.periodeendring.EndretFomDatoBekreftelse;
-import no.nav.k9.oppgave.bekreftelse.ung.periodeendring.EndretTomDatoBekreftelse;
+import no.nav.ung.kodeverk.behandling.BehandlingÅrsakType;
 import no.nav.ung.kodeverk.dokument.DokumentStatus;
-import no.nav.ung.kodeverk.ungdomsytelse.periodeendring.UngdomsprogramPeriodeEndringType;
+import no.nav.ung.kodeverk.etterlysning.EtterlysningStatus;
 import no.nav.ung.sak.behandlingslager.behandling.motattdokument.MottatteDokumentRepository;
-import no.nav.ung.sak.behandlingslager.behandling.startdato.UngdomsprogramBekreftetPeriodeEndring;
-import no.nav.ung.sak.behandlingslager.behandling.startdato.UngdomsytelseStartdatoRepository;
+import no.nav.ung.sak.behandlingslager.etterlysning.Etterlysning;
+import no.nav.ung.sak.behandlingslager.etterlysning.EtterlysningRepository;
 import no.nav.ung.sak.mottak.dokumentmottak.Trigger;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -22,42 +22,46 @@ import java.util.UUID;
 @OppgaveTypeRef(Bekreftelse.Type.UNG_ENDRET_TOM_DATO)
 public class DatoEndringBekreftelseHåndterer implements BekreftelseHåndterer {
 
-
-    private final UngdomsytelseStartdatoRepository ungdomsytelseStartdatoRepository;
     private final MottatteDokumentRepository mottatteDokumentRepository;
+    private final EtterlysningRepository etterlysningRepository;
+
     @Inject
-    public DatoEndringBekreftelseHåndterer(
-        UngdomsytelseStartdatoRepository ungdomsytelseStartdatoRepository, MottatteDokumentRepository mottatteDokumentRepository) {
-        this.ungdomsytelseStartdatoRepository = ungdomsytelseStartdatoRepository;
+    public DatoEndringBekreftelseHåndterer(MottatteDokumentRepository mottatteDokumentRepository, EtterlysningRepository etterlysningRepository) {
         this.mottatteDokumentRepository = mottatteDokumentRepository;
+        this.etterlysningRepository = etterlysningRepository;
     }
 
     @Override
     public void håndter(OppgaveBekreftelseInnhold oppgaveBekreftelse) {
         DatoEndring bekreftelse = oppgaveBekreftelse.oppgaveBekreftelse().getBekreftelse();
 
-        final var bekreftetPeriodeEndring = new UngdomsprogramBekreftetPeriodeEndring(
-            bekreftelse.getNyDato(),
-            oppgaveBekreftelse.mottattDokument().getJournalpostId(),
-            finnBekreftetPeriodeEndring(bekreftelse));
 
-        var behandling = oppgaveBekreftelse.behandling();
-        ungdomsytelseStartdatoRepository.lagre(behandling.getId(), bekreftetPeriodeEndring);
+        Etterlysning etterlysning = etterlysningRepository.hentEtterlysningForEksternReferanse(bekreftelse.getOppgaveId());
+
+        if (!etterlysning.getStatus().equals(EtterlysningStatus.VENTER)) {
+            throw new IllegalStateException("Etterlysning må hå status VENTER for å motta bekreftelse. Status var " + etterlysning.getStatus());
+        }
+
+        if (!bekreftelse.harBrukerGodtattEndringen()) {
+            Objects.requireNonNull(bekreftelse.getUttalelseFraBruker(),
+                "Uttalelsestekst fra bruker må være satt når bruker ikke har godtatt endringen");
+        }
+
+        etterlysning.mottattUttalelse(
+            oppgaveBekreftelse.mottattDokument().getJournalpostId(),
+            bekreftelse.harBrukerGodtattEndringen(),
+            bekreftelse.getUttalelseFraBruker()
+        );
+
+        etterlysningRepository.lagre(etterlysning);
         mottatteDokumentRepository.oppdaterStatus(List.of(oppgaveBekreftelse.mottattDokument()), DokumentStatus.GYLDIG);
     }
 
+
     @Override
     public Optional<Trigger> utledTrigger(UUID oppgaveId) {
-        return Optional.empty();
-    }
-
-    private static UngdomsprogramPeriodeEndringType finnBekreftetPeriodeEndring(DatoEndring bekreftelse) {
-        if (bekreftelse instanceof EndretTomDatoBekreftelse) {
-            return UngdomsprogramPeriodeEndringType.ENDRET_OPPHØRSDATO;
-        } else if (bekreftelse instanceof EndretFomDatoBekreftelse) {
-            return UngdomsprogramPeriodeEndringType.ENDRET_STARTDATO;
-        }
-        throw new IllegalArgumentException("Kunne ikke håndtere bekreftelse av type " + bekreftelse.getType());
+        Etterlysning etterlysning = etterlysningRepository.hentEtterlysningForEksternReferanse(oppgaveId);
+        return Optional.of(new Trigger(etterlysning.getPeriode(), BehandlingÅrsakType.UTTALELSE_FRA_BRUKER));
     }
 
 }
