+++ domenetjenester/ungdomsprogram/src/main/java/no/nav/ung/sak/ungdomsprogram/UngdomsprogramPeriodeTjeneste.java
@@ -85,7 +85,7 @@ public class UngdomsprogramPeriodeTjeneste {
     }
 
     public static LocalDateSegment<Boolean> erEndret(LocalDateInterval di, LocalDateSegment<Boolean> lhs, LocalDateSegment<Boolean> rhs) {
-        return new LocalDateSegment<>(di, lhs == null || rhs == null || lhs.getValue().equals(rhs.getValue()));
+        return new LocalDateSegment<>(di, lhs == null || rhs == null || !lhs.getValue().equals(rhs.getValue()));
     }
